+++ source/voxelman/gui/components.d
@@ -16,15 +16,16 @@ import voxelman.graphics : Color4ub;
 @Component("gui.WidgetTransform", Replication.none)
 struct WidgetTransform
 {
-	ivec2 relPos; // set by user
-	ivec2 size;   // set by user
-	ivec2 absPos; // updated in layout phase
-
+	ivec2 relPos; // set by user/layout
+	ivec2 size;   // set by user/layout
+	ivec2 minSize;// set by user
 	WidgetId parent; // 0 in root widget
 
-	ivec2 minSize; // updated in measure phase
-	void delegate(WidgetId) measureHandler; // called in measure phase
-	void delegate(WidgetId) layoutHandler;  // called in  layout phase
+	ivec2 absPos; // updated in layout phase from root to leaves by parent layout
+	ivec2 measuredSize; // updated in measure phase from leaves to root by widget's layout
+	void applyConstraints() { // applied by parent layout in measure phase
+		measuredSize = vector_max(measuredSize, minSize);
+	}
 }
 
 @Component("gui.WidgetStyle", Replication.none)
@@ -46,12 +47,55 @@ struct WidgetContainer
 	void put(WidgetId wId) {
 		children ~= wId;
 	}
+	void bringToFront(WidgetId wid) {
+		foreach(index, child; children)
+		{
+			if (child == wid)
+			{
+				moveItemToEnd(index, children);
+				return;
+			}
+		}
+	}
 }
 
-void bringToFront(GuiContext ctx, WidgetId widget)
+void moveItemToEnd(T)(size_t index, T[] array)
 {
-	auto parentId = ctx.widgets.get!WidgetTransform(widget).parent;
+	if (index+1 < array.length) // dont use length-1 due to size_t underflow
+	{
+		T item = array[index];
+		foreach(i; index..array.length-1)
+		{
+			array[i] = array[i+1];
+		}
+		array[$-1] = item;
+	}
+}
 
+unittest
+{
+	void test(T)(size_t index, T[] array, T[] expected) {
+		auto initial = array.dup;
+		moveItemToEnd(index, array);
+		import std.string : format;
+		assert(array == expected, format(
+			"for (%s, %s) expected %s, got %s",
+			index, initial, expected, array));
+	}
+	test(0, cast(int[])[], cast(int[])[]);
+	test(0, [3], [3]);
+	test(0, [3,2], [2,3]);
+	test(0, [3,2,1], [2,1,3]);
+	test(1, [3], [3]);
+	test(1, [3,2], [3,2]);
+	test(2, [3,2,1], [3,2,1]);
+}
+
+void bringToFront(GuiContext ctx, WidgetId wid)
+{
+	auto parentId = ctx.get!WidgetTransform(wid).parent;
+	if (auto container = ctx.get!WidgetContainer(parentId))
+		container.bringToFront(wid);
 }
 
 @Component("gui.WidgetRespondsToPointer", Replication.none)
@@ -63,11 +107,12 @@ struct WidgetIsFocusable {}
 @Component("gui.WidgetEvents", Replication.none)
 struct WidgetEvents
 {
+	import std.stdio;
 	this(Handlers...)(Handlers handlers)
 	{
 		addEventHandlers(handlers);
 	}
-	private alias EventHandler = bool delegate(WidgetId widgetId, ref void* event);
+	private alias EventHandler = void delegate(WidgetId wid, ref void* event);
 
 	@ignore EventHandler[][TypeInfo] eventHandlers;
 
@@ -100,18 +145,24 @@ struct WidgetEvents
 		eventHandlers.remove(typeid(T));
 	}
 
-	/// Returns true if event was handled
+	/// Returns true if any handlers were called
 	/// This handler will be called by Gui twice, before and after visiting its children.
 	/// In first case sinking flag will be true;
-	void postEvent(Event)(WidgetId widgetId, auto ref Event event)
+	bool postEvent(Event)(WidgetId wid, auto ref Event event)
 	{
-		foreach(handler; eventHandlers.get(typeid(Event), null))
+		if (auto handlers = typeid(Event) in eventHandlers)
 		{
-			auto payload = *cast(DelegatePayload*)&handler;
-			if (payload.contextPtr == FUNCTION_CONTEXT_VALUE)
-				(cast(void function(WidgetId, ref Event))payload.funcPtr)(widgetId, event);
-			else
-				(cast(void delegate(WidgetId, ref Event))handler)(widgetId, event);
+			foreach(handler; *handlers)
+			{
+				auto payload = *cast(DelegatePayload*)&handler;
+				if (payload.contextPtr == FUNCTION_CONTEXT_VALUE)
+					(cast(void function(WidgetId, ref Event))payload.funcPtr)(wid, event);
+				else
+					(cast(void delegate(WidgetId, ref Event))handler)(wid, event);
+			}
+			return true;
 		}
+		else
+			return false;
 	}
 }